+++ pkg/clustermanager/types.go
@@ -1,6 +1,7 @@
 package clustermanager
 
 import (
+	"github.com/F5Networks/k8s-bigip-ctlr/v2/config/client/clientset/versioned"
 	"k8s.io/client-go/kubernetes"
 )
 
@@ -22,7 +23,8 @@ type (
 	}
 
 	ClusterConfig struct {
-		KubeClient kubernetes.Interface
+		KubeClient   kubernetes.Interface
+		KubeCRClient versioned.Interface
 		// Maintain a mapping of managed resources -> services and services -> resources
 		//ManagedResources
 		// Maintain a mapping of managed informers -> services and services -> informers