+++ pkg/controller/resourceConfig_test.go
@@ -118,9 +118,32 @@ var _ = Describe("Resource Config Tests", func() {
 			name := formatCustomVirtualServerName("My_VS", 80)
 			Expect(name).To(Equal("My_VS_80"), "Invalid VirtualServer Name")
 		})
+		It("Pool name for TS", func() {
+			var name string
+			name = mockCtlr.formatPoolNameForTS(namespace, "svc1", intstr.IntOrString{IntVal: 80}, "app=test", "cluster1")
+			Expect(name).To(Equal("svc1_80_default_app_test"), "Invalid Pool Name for TS")
+			mockCtlr.multiClusterMode = PrimaryCIS
+			mockCtlr.discoveryMode = ""
+			name = mockCtlr.formatPoolNameForTS(namespace, "svc1", intstr.IntOrString{IntVal: 80}, "app=test", "cluster1")
+			Expect(name).To(Equal("ts_7bb616e6f5_multicluster"), "Invalid Pool Name for TS")
+			mockCtlr.discoveryMode = DefaultMode
+			name = mockCtlr.formatPoolNameForTS(namespace, "svc1", intstr.IntOrString{IntVal: 80}, "app=test", "cluster1")
+			Expect(name).To(Equal("ts_5bd74b770d_cluster1"), "Invalid Pool Name for TS")
+			mockCtlr.multiClusterMode = ""
+			mockCtlr.discoveryMode = ""
+		})
 		It("Pool Name", func() {
 			name := mockCtlr.formatPoolName(namespace, "svc1", intstr.IntOrString{IntVal: 80}, "app=test", "foo", "")
 			Expect(name).To(Equal("svc1_80_default_foo_app_test"), "Invalid Pool Name")
+		})
+		It("Monitor Name for TS", func() {
+			name := mockCtlr.formatMonitorNameForTS(namespace, "svc1", "http", intstr.IntOrString{IntVal: 80}, "foo.com", "path", "hash123")
+			Expect(name).To(Equal("svc1_default_foo_com_path_http_80"), "Invalid Monitor Name")
+			mockCtlr.multiClusterMode = PrimaryCIS
+			name = mockCtlr.formatMonitorNameForTS(namespace, "svc1", "http", intstr.IntOrString{IntVal: 80}, "foo.com", "path", "hash123")
+			Expect(name).To(Equal("ts_hash123_80_http"), "Invalid Pool Name for TS")
+			mockCtlr.multiClusterMode = ""
+			mockCtlr.discoveryMode = ""
 		})
 		It("Monitor Name", func() {
 			name := formatMonitorName(namespace, "svc1", "http", intstr.IntOrString{IntVal: 80}, "foo.com", "path")
@@ -712,7 +735,7 @@ var _ = Describe("Resource Config Tests", func() {
 			svc.Annotations = make(map[string]string)
 			svc.Annotations[HealthMonitorAnnotation] = `{"interval": 5, "timeout": 10}`
 
-			err := mockCtlr.prepareRSConfigFromLBService(rsCfg, svc, svcPort)
+			err := mockCtlr.prepareRSConfigFromLBService(rsCfg, svc, svcPort, "")
 			Expect(err).To(BeNil(), "Failed to Prepare Resource Config from Service")
 			Expect(len(rsCfg.Pools)).To(Equal(1), "Failed to Prepare Resource Config from Service")
 			Expect(len(rsCfg.Monitors)).To(Equal(1), "Failed to Prepare Resource Config from Service")